+++ libs/engine/src/vivarium/engine/framework/logging/__init__.py
@@ -8,7 +8,9 @@ Logging
 from vivarium.engine.framework.logging.interface import LoggingInterface
 from vivarium.engine.framework.logging.manager import LoggingManager
 from vivarium.engine.framework.logging.utilities import (
+    add_logging_sink,
     configure_logging_to_file,
     configure_logging_to_terminal,
+    get_log_level,
     list_loggers,
 )