+++ libs/engine/src/vivarium/engine/interface/cli.py
@@ -42,6 +42,7 @@ from vivarium.engine.framework.logging import (
     configure_logging_to_terminal,
 )
 from vivarium.engine.framework.utilities import handle_exceptions
+from vivarium.engine.interface.cli_tools import verbose_option
 from vivarium.engine.interface.utilities import get_output_root
 
 
@@ -74,18 +75,7 @@ def simulate() -> None:
     help="The directory to write results to. A folder will be created "
     "in this directory with the same name as the configuration file.",
 )
-@click.option(
-    "--verbose",
-    "-v",
-    is_flag=True,
-    help="Logs verbosely. Useful for debugging and development.",
-)
-@click.option(
-    "--quiet",
-    "-q",
-    is_flag=True,
-    help="Suppresses all logging except for warnings and errors.",
-)
+@verbose_option()
 @click.option(
     "--pdb",
     "with_debugger",
@@ -96,8 +86,7 @@ def run(
     model_specification: Path,
     artifact_path: Path,
     results_directory: Path,
-    verbose: bool,
-    quiet: bool,
+    verbose: int,
     with_debugger: bool,
 ) -> None:
     """Run a simulation from the command line.
@@ -109,10 +98,7 @@ def run(
     MODEL_SPECIFICATION if one does not exist. Results will be written to a
     further subdirectory named after the start time of the simulation run.
     """
-    if verbose and quiet:
-        raise click.UsageError("Cannot be both verbose and quiet.")
-    verbosity = 1 + int(verbose) - int(quiet)
-    configure_logging_to_terminal(verbosity=verbosity, long_format=False)
+    configure_logging_to_terminal(verbosity=verbose, long_format=False)
 
     start = time()